Sat, Sep 7, 2024, 12:28 PM +05:30

+++ sentimental-credit/credit.py
@@ -8,7 +8,7 @@ while True:
         remainder = number % 100
         left = 2 * (remainder / 10)
         if left > 9:
-            ctr = left/10 + left % 10
+            ctr = int(left/10) + left % 10
         else:
             ctr = ctr + left
